server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
httpx>=0.27.0
pdfplumber>=0.11.0
pymupdf>=1.24.0
tzdata>=2024.1
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.9.0
email-validator>=2.0.0
